fix: object mistaken problem when dealing identity cards
Identity cards' number and types were mistaken in three-player games, and the setup procedures were interrupted, resulting in no identity cards being dealt. The issue seems to occur in the function `transferObjectsWithSameTag`. This function checks if cards in the identity deck do not have specific tags (indicating whether the card is needed for the current player count); if not, the function removes the card by its index.

The mistake might be attributed to the iterative behavior that continuously removes objects and updates the index, causing the taken object not to match the compared one. The loop interrupts when the index of nothing (nil) is reached.

To address this problem, the object's GUID is now used instead of the index. Switching to the object's GUID ensures a unique and consistent way to refer to objects, eliminating the risk of misalignment.

For further details on the method, please refer to [this link](https://www.reddit.com/r/tabletopsimulator/comments/e8yp68/scripting_question_hot_to_get_cards_with_the_same/). Instructions in the official API tutorial did not work.

wy-8