feat: script zones for flexible deck operation
script zones were added to increase deck operation flexibility. For example, When a deck gets down to one card, it is considered a new object - a single card. This card can not be accessed by deck operations (as the deck was destructed).

A workaround is using script zone to get the remained card, and doing
what you want to do.

wy-8