fix: clean remaining identity cards for 3-player setup
In a 3-player setup, 4 identity cards are prepared and distributed. When the deck gets down to one card, it's destroyed, and the remaining card cannot be accessed with operations based on the deck GUID.

The workaround is to use a script zone, waiting for the other cards to leave the area first, then identifying and moving the cards in the script zone to a specified location or container. Refer to the strategy at [Reddit](https://www.reddit.com/r/tabletopsimulator/comments/lpgbl3/drawing_last_card_from_deck/).

wy-8